--- a/src/tests/class_tests/openms/source/CompNovoIonScoringCID_test.cpp
+++ b/src/tests/class_tests/openms/source/CompNovoIonScoringCID_test.cpp
@@ -82,7 +82,7 @@ START_SECTION((void scoreSpectrum(Map<double, IonScore>& CID_ion_scores, PeakSpe
   TheoreticalSpectrumGenerator tsg;
   Param tsg_param(tsg.getParameters());
   tsg_param.setValue("add_losses", "true");
-  tsg_param.setValue("add_isotopes", "true");
+  tsg_param.setValue("isotope_model", "coarse");
   tsg.setParameters(tsg_param);
 
   PeakSpectrum rspec;