… src/tests/class_tests/openms/source/CoarseIsotopeDistribution_test.cpp
… src/tests/class_tests/openms/source/CompNovoIdentificationCID_test.cpp
… src/tests/class_tests/openms/source/CompNovoIonScoring_test.cpp
… src/tests/class_tests/openms/source/Deisotoper_test.cpp
… src/tests/class_tests/openms/source/FineIsotopeDistribution_test.cpp
… src/tests/class_tests/openms/source/TheoreticalSpectrumGenerator_test.cpp