Work on TheoreticalSpectrumGenerator

- allow choice of isotopic generator (coarse / fine)
- speedup if no sorting is needed
- general code improvements
- substantial speedup for loss calculation (fast path)

# Benchmarking

Benchmark on peptide `DFPIANGER`, all measurements are against Release 2.4 (pyopenms). Benchmark code:

```  
  auto tsg= TheoreticalSpectrumGenerator();
  auto pep = AASequence::fromString("DFPIANGER");
  auto spec = MSSpectrum();
  
  for (Size i = 0; i < nr_mol;  i++) 
  {
    tsg.getSpectrum(spec, pep, 1,2);
    nr_fragments += spec.size();
    spec.clear(true);
  }
```

where `false` was used in the last line if no annotations are present.

## Single y ion series calculation

before: 1.2 usec / loop
after: 0.25 usec / loop (no sorting)
after: 0.71 usec / loop (with sorting)

single ion series calculations are ca 5x faster which is due code improvements, turning off sorting and to removing the `shrink_to_fit` from the `clear()`. Pure code improvements in TSG alone are about 50% faster (see below).

### Single y ion series calculation (TSG only)

Using `spec.clear(true);` forcing the library to clear all arrays and `shrink_to_fit` all arrays and turning on sorting, we get

before: 1.5 usec / loop
after: 0.97 usec / loop

## Single y ion series calculation with losses

before: 429 usec / loop
after: 4.9 usec / loop (no sorting)

This is mainly due to calls of `EmpiricalFormula::getMonoWeight()` since the losses are coded to individual residues and the weight of the loss needs to be re-computed each time.  This is still substantially faster than before.

## Single y ion series calculation with losses and annotation

before: 528 usec / loop
after: 34 usec / loop (no sorting)

This is mainly due to calls of `EmpiricalFormula::toString()` which is pretty slow as well as storing the annotations and dealing with Strings in general.

## Single y ion series calculation with losses, annotation and isotopes (coarse)

before: 1040 usec / loop
after: 640 usec / loop (no sorting)

## Single y ion series calculation with losses, annotation and isotopes (fine)

before: NA 
after: 1560 usec / loop (no sorting)