@timosachsenberg 2020-01-05
Some nice additions and optimizations. Maybe @enetz can also provide some insights on e.g. sorting .