@hroest 2020-01-09
@enetz I think there may be an easy way to add a simple isotopic model to the TSG, basically one that does not really take intensity into account but simply adds a peak at a given intensity and C12-C13 carbon mass away. I think doing that could actually reduce the execution time from ms to usec, speeding up TSG at least 100 or even more fold. What do you think, would that be interesting?