@enetz 2020-01-10
@hroest in my very specialized TSG for XL-MS I already do that. Because I deisotope anything with 3 or more peaks in the experimental spectra anyway, I just add a second peak for each fragment like this:
```
        spectrum.emplace_back(pos, charge);
        if (add_isotopes_)
        {
          spectrum.emplace_back(pos+(Constants::C13C12_MASSDIFF_U / charge), charge);
        }
```
This could also be generalized for higher numbers of peaks with a simple loop. I think most of our tools do not use the intensity of theoretical peaks anyway.