@enetz 2020-01-10
About sorting, in my tool the sorting of theoretical spectra is still a big chunk of total runtime.
I tested several sorting algorithms and boost::pdqsort turned out to be the most efficient for the two types of spectra that OpenPepXL generates. E.g. creating 2 million linear ion spectra (the type most similar to normal peptide spectra) took 23s with std::sort and 18s with boost::pdqsort.  Also to what extent the spectra are already partially sorted can have an influence on sorting time as well, so if possible it helps to add peaks in order as much as possible, e.g. first add neutral loss peaks, then monoisotopic, then additional isotopic pattern peaks.