@timosachsenberg 2020-01-13
ok I think this is good to merge as is... any more comments on the neutral loss? my take is to remove it if we can't find out the reason why its there