@cbielow 2020-01-20
>Also to what extent the spectra are already partially sorted can have an influence on sorting time as 
>well, so if possible it helps to add peaks in order as much as possible, e.g. first add neutral loss 
>peaks, then monoisotopic, then additional isotopic pattern peaks.

have you tried storing the each type of loss in separate vectors and use https://en.cppreference.com/w/cpp/algorithm/merge on the result? (C17 also has https://en.cppreference.com/w/cpp/algorithm/inplace_merge but alas...)